cms_project/requirements.txt
Django==4.2.5
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.6.0
mysql==0.0.3
mysqlclient==2.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
